fix: apply autograder suggestions

+++ controllers/authController.js
@@ -9,17 +9,19 @@ import jwt from 'jsonwebtoken';
 const JWT_SECRET = process.env.JWT_SECRET;
 const JWT_EXPIRES = '1h';
 
-const newUserSchema = z.object({
-  nome: z.string("O campo 'nome' deve ser uma string.").min(1, "O campo 'nome' é obrigatório."),
-  email: z.email("O campo 'email' deve ser um email válido").nonempty("O campo 'email' é obrigatório."),
-  senha: z
-    .string("O campo 'senha' deve ser uma string.")
-    .min(8, 'A senha deve ter pelo menos 8 caracteres.')
-    .regex(/[a-z]/, 'A senha deve conter pelo menos uma letra minúscula.')
-    .regex(/[A-Z]/, 'A senha deve conter pelo menos uma letra maiúscula.')
-    .regex(/[0-9]/, 'A senha deve conter pelo menos um número.')
-    .regex(/[^A-Za-z0-9]/, 'A senha deve conter pelo menos um caractere especial.'),
-});
+const newUserSchema = z
+  .object({
+    nome: z.string("O campo 'nome' deve ser uma string.").min(1, "O campo 'nome' é obrigatório."),
+    email: z.email("O campo 'email' deve ser um email válido").nonempty("O campo 'email' é obrigatório."),
+    senha: z
+      .string("O campo 'senha' deve ser uma string.")
+      .min(8, 'A senha deve ter pelo menos 8 caracteres.')
+      .regex(/[a-z]/, 'A senha deve conter pelo menos uma letra minúscula.')
+      .regex(/[A-Z]/, 'A senha deve conter pelo menos uma letra maiúscula.')
+      .regex(/[0-9]/, 'A senha deve conter pelo menos um número.')
+      .regex(/[^A-Za-z0-9]/, 'A senha deve conter pelo menos um caractere especial.'),
+  })
+  .strict();
 
 const loginSchema = z.object({
   email: z.email("O campo 'email' deve ser um email válido").nonempty("O campo 'email' é obrigatório."),
@@ -79,7 +81,7 @@ async function login(req, res, next) {
   }
 }
 
-async function logout(req, res) {
+async function logout(req, res, next) {
   try {
     const authHeader = req.headers['authorization'];
     const token = authHeader?.split(' ')[1];
@@ -87,26 +89,30 @@ async function logout(req, res) {
     if (!token) {
       return next(createError(401, 'Token não fornecido.'));
     }
-    
+
     blacklist.add(token);
 
     return res.status(200).send();
   } catch (err) {
-        if (err.name === 'ZodError') {
+    if (err.name === 'ZodError') {
       return next(createError(400, formatZodErrors(err)));
     }
     return next(err);
   }
 }
 
-async function deleteUser(req, res) {
+async function deleteUser(req, res, next) {
   try {
-    const userId = req.params.id;
+    const userId = req.params.id ? Number(req.params.id) : null;
+
+    if (!userId) {
+      return next(createError(400, 'ID inválido.'));
+    }
 
     const usuario = await usuariosRepository.find(userId);
 
     if (!usuario) {
-      return next(createError(400, 'Usuário não encontrado.'));
+      return next(createError(404, 'Usuário não encontrado.'));
     }
 
     await usuariosRepository.remove(usuario.id);
